+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 
 
-## [0.0.6] - 2025-2-10
+## [0.0.6]
 ### 1. Add parameter cardWidth, that for get card width for do flexible the titles and subtitles that inside positions to make it responsive .
 ### 2. Delete 'sar' form title, for show just title without show RichText  'sar' .
 ### 3. Use LayoutBuilder with CardListView and CardGridView to get maxWidth size for add limitation for position to can flex the element inside it .  
 ### 4. Show showSnackBar on pressed to onPressedFavoriteIcon
+
+
+## [0.0.7]
+### 1. Create DetailsScreen and make user onTap to card Navigate to it.
+### 2. Fix bug onClick to Favorite icon.
+### 3. Delete custom_local_hero code Example 
+### 4. Add example to how to add custom image  
+
+